Split temperature and density diagnostics into their own file
