+++ spaxelsleuth/utils/linefns.py
@@ -7,6 +7,7 @@ from spaxelsleuth.utils.misc import remove_col_suffix, add_col_suffix
 import logging
 logger = logging.getLogger(__name__)
 
+
 bpt_dict = {
      "0.0": "SF",
      "1.0": "Composite",
@@ -15,10 +16,12 @@ bpt_dict = {
      "4.0": "Ambiguous",
      "-1.0": "Not classified",
 }
+
+
 def bpt_num_to_str(s):
     return [bpt_dict[str(a)] for a in s]
 
-######################################################################
+
 def compute_eline_luminosity(df, ncomponents_max, eline_list, flux_units):
     """Compute emission line luminosities, where the emission line flux units are given by flux_units erg s^-2 cm^-2."""
     logger.debug(f"computing emission line luminosities...")
@@ -35,7 +38,7 @@ def compute_eline_luminosity(df, ncomponents_max, eline_list, flux_units):
 
     return df
 
-######################################################################
+
 def compute_FWHM(df, ncomponents_max):
     """Compute the Full-Width at Half Maximum from the velocity dispersion."""
     logger.debug(f"computing emission line FWHMs...")
@@ -48,7 +51,6 @@ def compute_FWHM(df, ncomponents_max):
     return df
 
 
-# Reference lines from literature
 def Kewley2001(ratio_x, ratio_x_vals, log=True):
     """
     Returns the "maximum starburst" demarcation lines of Kewley et al. (2001)
@@ -88,6 +90,7 @@ def Kewley2001(ratio_x, ratio_x_vals, log=True):
         ratio_x_vals[ratio_x_vals > -0.59] = np.nan
         return 0.73 / (ratio_x_vals + 0.59) + 1.33
 
+
 def Kauffman2003(ratio_x, ratio_x_vals, log=True):
     """
     Returns the empirically derived star-forming/AGN demarcation line of 
@@ -120,6 +123,7 @@ def Kauffman2003(ratio_x, ratio_x_vals, log=True):
     if ratio_x == "log N2":
         ratio_x_vals[ratio_x_vals > 0.05] = np.nan
         return 0.61 / (ratio_x_vals - 0.05) + 1.3
+
 
 def Kewley2006(ratio_x, ratio_x_vals, log=True):
     """
@@ -156,6 +160,7 @@ def Kewley2006(ratio_x, ratio_x_vals, log=True):
     elif ratio_x == "log O1":
         ratio_x_vals[ratio_x_vals < -1.1259] = np.nan
         return 1.18 * ratio_x_vals + 1.30
+
 
 def Law2021_1sigma(ratio_x, ratio_x_vals, log=True):
     """
@@ -231,422 +236,6 @@ def Law2021_3sigma(ratio_x, ratio_y_vals, log=True):
     elif ratio_x == "log O1":
         ratio_y_vals[ratio_y_vals > 0.65] = np.nan
         return 18.664 * ratio_y_vals**4 - 36.343 * ratio_y_vals**3 + 22.238 * ratio_y_vals**2 - 6.134 * ratio_y_vals - 0.283
-
-
-def get_n_e_Proxauf2014(R):
-    """
-    Electron density computation from eqn. 3 of Proxauf et al. (2014), which 
-    assumes an electron temperature of 10^4 K.
-
-    INPUT
-    --------------------------------------------------------------------------
-    R:          float
-        Flux ratio F([SII]6716) / F([SII]6731).
-
-    OUTPUT
-    --------------------------------------------------------------------------
-    tuple of 
-        (n_e, lolim_mask, uplim_mask)
-    where n_e is the electron density (in cm^-3) computed using eqn. 3 of 
-    Proxauf (2014), and lolim_mask and uplim_mask areboolean arrays the 
-    same dimensions as n_e corresponding to indices where the line ratio 
-    corresponds to electron density lower and upper limits of 40 cm^-3 and 
-    10^4 cm^-3 respectively.
-
-    REFERENCES
-    --------------------------------------------------------------------------  
-    https://ui.adsabs.harvard.edu/abs/2014A%26A...561A..10P/abstract
-
-    """
-    log_n_e = 0.0543 * np.tan(-3.0553 * R + 2.8506)\
-             + 6.98 - 10.6905 * R\
-             + 9.9186 * R**2 - 3.5442 * R**3
-    n_e = 10**log_n_e
-
-    # High & low density limits
-    lolim_mask = n_e < 40
-    uplim_mask = n_e > 1e4
-    n_e[lolim_mask] = 40
-    n_e[uplim_mask] = 1e4
-
-    return n_e, lolim_mask, uplim_mask
-
-
-def get_n_e_Sanders2016(ratio, R):
-    """
-    Electron density computation from eqn. 7 of Sanders et al. (2016), which 
-    assumes an electron temperature of 10^4 K.
-
-    INPUT
-    --------------------------------------------------------------------------
-    ratio:      str
-        Which emission line to use in the diagnostic. Options are "[SII]" and 
-        "[OII]".
-    
-    R:          float
-        Flux ratio. Must correspond to 'ratio':
-            [SII]: F([SII]6716) / F([SII]6731) 
-            [OII]: F([OII]3729) / F([OII]3726) 
-
-    OUTPUT
-    --------------------------------------------------------------------------
-    tuple of 
-        (n_e, lolim_mask, uplim_mask)
-    where n_e is the electron density (in cm^-3) computed using eqn. 7 of 
-    Sanders et al. (2016), and lolim_mask and uplim_mask areboolean arrays the 
-    same dimensions as n_e corresponding to indices where the line ratio 
-    corresponds to electron density lower and upper limits of 1 cm^-3 and 
-    10^5 cm^-3 respectively.
-
-    REFERENCES
-    --------------------------------------------------------------------------  
-    https://ui.adsabs.harvard.edu/abs/2016ApJ...816...23S/abstract 
-
-    """
-    # Repalce infs with NaNs
-    R = R.copy()
-    R[np.isinf(R)] = np.nan
-
-    # Coefficients 
-    if ratio == "[SII]":
-        a = 0.4315
-        b = 2107
-        c = 627.1
-        R_min = 0.4375
-        R_max = 1.4484
-    elif ratio == "[OII]":
-        a = 0.3771
-        b = 2468
-        c = 638.4
-        R_min = 0.3839
-        R_max = 1.4558
-    else:
-        raise ValueError(f"{ratio} is not a valid line ratio! Valid ratios are [OII] and [SII]!")
-
-    # Calculate the electron density 
-    n_e = (c * R - a * b) / (a - R)
-    
-    # Apply upper/lower limits
-    n_e_min = 1
-    n_e_max = 1e5
-    lolim_mask = R > R_max
-    uplim_mask = R < R_min
-    n_e[lolim_mask] = n_e_min
-    n_e[uplim_mask] = n_e_max
-    
-    return n_e, lolim_mask, uplim_mask
-
-
-def compute_electron_density(df, diagnostic, ratio, s=None):
-    """
-    Calculate electron densities using emission line ratios of [SII] or [OII].
-
-    INPUTS
-    --------------------------------------------------------------------------
-    df:         pandas DataFrame
-        DataFrame in which to compute 
-
-    diagnostic: str
-        Which diagnostic to use. Options are "Proxauf2014" and "Sanders2016".
-
-    ratio:      str
-        Which emission line to use in the diagnostic. Options are "[SII]" 
-        (valid for both "Proxauf2014" and "Sanders2016") and "[OII]" 
-        ("Sanders2016" only).
-
-    s:          str 
-        Column suffix to trim before carrying out computation - e.g. if 
-        you want to compute metallicities for "total" fluxes, and the 
-        columns of the DataFrame look like 
-
-            "HALPHA (total)", "HALPHA error (total)", etc.,
-
-        then setting s=" (total)" will mean that this function "sees"
-
-            "HALPHA", "HALPHA error".
-
-        Useful for running this function on different emission line 
-        components. The suffix is added back to the columns (and appended
-        to any new columns that are added) before being returned. For 
-        example, using the above example, the new added columns will be 
-
-            "log N2 (total)", "log N2 error (lower) (total)", 
-            "log N2 error (upper) (total)"
-
-    OUTPUTS
-    -----------------------------------------------------------------------
-    The original DataFrame with new columns added.
-
-    REFERENCES
-    --------------------------------------------------------------------------  
-    https://ui.adsabs.harvard.edu/abs/2014A%26A...561A..10P/abstract
-    https://ui.adsabs.harvard.edu/abs/2016ApJ...816...23S/abstract 
-    """
-    # Trim suffix
-    df, suffix_cols, suffix_removed_cols, old_cols = remove_col_suffix(df, s)
-    old_cols = df.columns
-
-    if (ratio == "[SII]" and "[SII] ratio" in df) or (ratio == "[OII]" and "[OII] ratio" in df):
-
-        # Compute diagnostics 
-        if diagnostic == "Proxauf2014":
-            if ratio != "[SII]":
-                raise ValueError(f"Invalid ratio {ratio} for diagnostic {diagnostic}!")
-            R = df["[SII] ratio"].values
-            n_e, cond_lower_lim, cond_upper_lim = get_n_e_Proxauf2014(R)
-        elif diagnostic == "Sanders2016":
-            if ratio not in["[SII]", "[OII]"]:
-                raise ValueError(f"Invalid ratio {ratio} for diagnostic {diagnostic}!")
-            R = df[f"{ratio} ratio"].values
-            n_e, cond_lower_lim, cond_upper_lim = get_n_e_Sanders2016(ratio, R)
-        else:
-            raise ValueError(f"{diagnostic} is an invalid diagnostic!")
-
-        df[f"n_e ({diagnostic} ({ratio}))"] = n_e
-        df[f"n_e saturation flag ({diagnostic} ({ratio}))"] = False
-        df.loc[cond_lower_lim | cond_upper_lim, f"n_e saturation flag ({diagnostic} ({ratio}))"] = True
-    
-    else:
-        if ratio == "[SII]":
-            logger.warning(f"not computing n_e because I could not find the column '[SII] ratio'")
-        elif ratio == "[OII]":
-            logger.warning(f"not computing n_e because I could not find the column '[OII] ratio'")
-
-    # Rename columns
-    df = add_col_suffix(df, s, suffix_cols, suffix_removed_cols, old_cols)
-
-    return df
-
-
-def get_T_e_Proxauf2014(R):
-    """"
-    Electron temperature computation from eqn. 1 of Proxauf et al. (2014).
-    Valid for electron densitites up to a few 10^4 cm^-3. 
-
-    INPUT
-    --------------------------------------------------------------------------
-    R:          float
-        Flux ratio F([OIII]4959,5007) / F([OIII]4363).
-
-    OUTPUT
-    --------------------------------------------------------------------------
-    tuple of 
-        (T_e, lolim_mask, uplim_mask)
-    where T_e is the electron temperature (in K) computed using eqn. 1 of 
-    Proxauf (2014), and lolim_mask and uplim_mask areboolean arrays the 
-    same dimensions as n_e corresponding to indices where the line ratio 
-    corresponds to electron temperature lower and upper limits of 5000 K and 
-    24000 K respectively.
-
-    REFERENCES
-    --------------------------------------------------------------------------  
-    https://ui.adsabs.harvard.edu/abs/2014A%26A...561A..10P/abstract
-    """
-    # Repalce infs with NaNs
-    R = R.copy()
-    R[np.isinf(R)] = np.nan
-
-    # Calculate the electron temperature
-    r = np.log10(R)
-    T_e = 5294 * (r - 0.848)**(-1) + 19047 - 7769 * r + 944 * r**2
-
-    # Apply upper/lower limits
-    r_min = 1.25
-    r_max = 3.75
-    T_e_min = 5e3
-    T_e_max = 24e3
-    lolim_mask = (r > r_max)
-    uplim_mask = (r < r_min)
-    T_e[lolim_mask] = T_e_min
-    T_e[uplim_mask] = T_e_max            
-
-    return T_e, lolim_mask, uplim_mask
-
-
-def get_T_e_PM2014(ratio, R):
-    """"
-    Electron temperature computation from eqn. 1 of Perez Montero (2014).
-    Assumes a density of 100 cm^-3. Note that the returned temperatures 
-    reflect the temperature associated with the ions corresponding to the 
-    chosen line ratios. 
-
-    INPUT
-    --------------------------------------------------------------------------
-    ratio:      str
-        Which emission line to use in the diagnostic. Options are "[NII]" and
-        "[OIII]". 
-        
-        NOTE: the [SIII] diagnostic included in this paper has not 
-        been implemented due to a suspected typo in their eqn. 11; specifying
-        the [SIII] ratio will raise a NotImplementedError. 
-    
-    R:          float
-        Flux ratio. Must correspond to 'ratio':
-            [NII]: F([NII]6548,6584) / F([NII]5755) 
-            [OII]: F([OIII]4959,5007) / F([OIII]4363) 
-            [SIII]: F([SIII]9069,9531) / F([SIII]6312) 
-
-    OUTPUT
-    --------------------------------------------------------------------------
-    tuple of 
-        (T_e, lolim_mask, uplim_mask)
-    where T_e is the electron temperature (in K) computed using eqns. 5 and 7 
-    of Perez Montero (2014), and lolim_mask and uplim_mask are boolean arrays 
-    the same dimensions as n_e corresponding to indices where the line ratio 
-    corresponds to valid electron temperatures which are in the range 
-        * (0.7, 2.5) * 10^4 K for the [OIII] diagnostic,
-        * (0.6, 2.2) * 10^4 K for the [NII] diagnostic, and 
-        * (0.6, 2.5) * 10^4 K for the [SIII] diagnostic.
-
-    REFERENCES
-    --------------------------------------------------------------------------  
-    https://ui.adsabs.harvard.edu/abs/2014MNRAS.441.2663P/abstract
-    """
-    # Repalce infs with NaNs
-    R = R.copy()
-    R[np.isinf(R)] = np.nan
-
-    # Calculate the electron temperature
-    if ratio == "[OIII]":
-        T_e = (0.7840 - 0.0001357 * R + 48.44 / R) * 1e4
-        T_e_min = 0.7e4
-        T_e_max = 2.5e4
-    elif ratio == "[NII]":
-        T_e = (0.6153 - 0.0001529 * R + 35.3641 / R) * 1e4
-        T_e_min = 0.6e4
-        T_e_max = 2.2e4
-    elif ratio == "[SIII]":
-        raise NotImplementedError("For now, the [SIII] diagnostic has not been implemented due to a suspected typo in their eqn. 11")
-        """
-        If you plot the below equation, which is as printed in the paper (eqn. 11),
-        as a function of R, you will see that it has a minimum at approx. R = 300
-        and is therefore two-valued within the specified range of valid temperatures.
-        I suspect that the plus sign before 0.0003187 is actually meant to be a minus
-        sign, which would make it consistent with the other two diagnostics. Indeed,
-        if it is replaced with a minus sign the minimum disappears and the values look
-        more reasonable. 
-        """
-        T_e = (0.5147 + 0.0003187 * R + 23.64041 / R) * 1e4
-        T_e_min = 0.6e4
-        T_e_max = 2.5e4
-
-    else:
-        raise ValueError(f"{ratio} is not a valid line ratio! Valid ratios are [OIII] and [NII]!")
-
-    lolim_mask = (T_e < T_e_min)
-    uplim_mask = (T_e > T_e_max)
-    T_e[lolim_mask] = T_e_min
-    T_e[uplim_mask] = T_e_max            
-
-    return T_e, lolim_mask, uplim_mask
-
-
-def compute_electron_temperature(df, diagnostic, ratio="[OIII]", s=None):
-    """
-    Calculate electron temperature using emission line ratios of [OIII].
-
-    INPUTS
-    --------------------------------------------------------------------------
-    df:         pandas DataFrame
-        DataFrame in which to compute the electron temperature.
-
-    diagnostic: str
-        Which diagnostic to use. Options are "Proxauf2014" and "PM2014".
-
-    ratio:      str
-        Which emission line to use in the diagnostic. Valid options are
-        * "[OIII]" for the "Proxauf2014" diagnostic or 
-        * "[OIII]", "[NII]" or "[SIII]" for the "PM2014" diagnostic.        
-
-    s:          str 
-        Column suffix to trim before carrying out computation - e.g. if 
-        you want to compute metallicities for "total" fluxes, and the 
-        columns of the DataFrame look like 
-
-            "HALPHA (total)", "HALPHA error (total)", etc.,
-
-        then setting s=" (total)" will mean that this function "sees"
-
-            "HALPHA", "HALPHA error".
-
-        Useful for running this function on different emission line 
-        components. The suffix is added back to the columns (and appended
-        to any new columns that are added) before being returned. For 
-        example, using the above example, the new added columns will be 
-
-            "T_e (<diagnostic>) (<ratio>) (total)" and
-            "T_e saturation flag (<diagnostic>) (<ratio>) (total)"
-
-    OUTPUTS
-    -----------------------------------------------------------------------
-    The original DataFrame with new columns added:
-
-        "T_e (<diagnostic>) (<ratio>) <suffix>": electron temperature in K
-        
-        "T_e saturation flag (<diagnostic>) (<ratio>) <suffix>": True if 
-        the line ratios correspond to "saturation" in the diagnostic used, 
-        False otherwise. T_e measurements corresponding to "saturated" line 
-        ratios will be set to the upper/lower limits for the diagnostic.
-
-    REFERENCES
-    --------------------------------------------------------------------------  
-    https://ui.adsabs.harvard.edu/abs/2014A%26A...561A..10P/abstract
-    https://ui.adsabs.harvard.edu/abs/2016ApJ...816...23S/abstract 
-    """
-    
-    # Input checking 
-    if diagnostic not in ["Proxauf2014", "PM2014"]:
-        raise ValueError(f"{diagnostic} is an invalid diagnostic!")
-    if diagnostic == "Proxauf2014":
-        if ratio not in ["[OIII]"]:
-            raise ValueError(f"Line ratio {ratio} is invalid for the {diagnostic} T_e diagnostic!")
-    elif diagnostic == "PM2014":
-        if ratio not in ["[OIII]", "[NII]", "[SIII]"]:
-            raise ValueError(f"Line ratio {ratio} is invalid for the {diagnostic} T_e diagnostic!")    
-    
-    # Trim suffix
-    df, suffix_cols, suffix_removed_cols, old_cols = remove_col_suffix(df, s)
-    old_cols = df.columns
-
-    # Calculate the electron temperature 
-    if ratio == "[OIII]":
-        if f"OIII4363" not in df and "OIII4959+OIII5007" not in df:
-            logger.warning(f"not computing T_e because I could not find the columns OIII4363 and/or OIII4959+OIII5007")
-        else:
-            # Calculate 
-            R = df["OIII4959+OIII5007"] / df[f"OIII4363"]
-            if diagnostic == "Proxauf2014":
-                T_e, cond_lower_lim, cond_upper_lim = get_T_e_Proxauf2014(R)
-            elif diagnostic == "PM2014":
-                T_e, cond_lower_lim, cond_upper_lim = get_T_e_PM2014(ratio="[OIII]", R=R)
-
-    elif ratio == "[NII]":
-        if "NII6548+NII6583" not in df and "NII5755" not in df:
-            logger.warning(f"not computing T_e because I could not find the columns NII5755 and/or NII6548+NII6583")
-        else:
-            # Calculate 
-            R = df["NII6548+NII6583"] / df[f"NII5755"]
-            if diagnostic == "PM2014":
-                T_e, cond_lower_lim, cond_upper_lim = get_T_e_PM2014(ratio="[NII]", R=R)
-
-    elif ratio == "[SIII]":
-        if "SIII9069+SIII9531" not in df and "SIII6312" not in df:
-            logger.warning(f"not computing T_e because I could not find the columns SIII6312 and/or SIII9069+SIII9531")
-        else:
-            # Calculate 
-            R = df["SIII9069+SIII9531"] / df[f"SIII6312"]
-            if diagnostic == "PM2014":
-                T_e, cond_lower_lim, cond_upper_lim = get_T_e_PM2014(ratio="[SIII]", R=R)
-
-    # Store in DataFrame
-    df[f"T_e ({diagnostic} ({ratio}))"] = T_e
-    df[f"T_e saturation flag ({diagnostic} ({ratio}))"] = False
-    df.loc[cond_lower_lim | cond_upper_lim, f"T_e saturation flag ({diagnostic} ({ratio}))"] = True
-
-    # Rename columns
-    df = add_col_suffix(df, s, suffix_cols, suffix_removed_cols, old_cols)
-
-    return df
 
 
 def bpt_fn(df, s=None):
@@ -797,7 +386,6 @@ def bpt_fn(df, s=None):
     df = add_col_suffix(df, s, suffix_cols, suffix_removed_cols, old_cols)
 
     return df
-
 
 
 def law2021_fn(df, s=None):
